@@ -1,34 +1,20 @@
 import React, { useContext, useEffect, useState } from 'react';
-import Container from '@material-ui/core/Container';
-import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core';
-import Grid from '@material-ui/core/Grid';
-import CardActionArea from '@material-ui/core/CardActionArea';
-import CardMedia from '@material-ui/core/CardMedia';
-import CardContent from '@material-ui/core/CardContent';
-import Card from '@material-ui/core/Card';
+import Container from '@mui/material/Container';
+import Typography from '@mui/material/Typography';
+import Grid from '@mui/material/Grid';
+import CardActionArea from '@mui/material/CardActionArea';
+import CardMedia from '@mui/material/CardMedia';
+import CardContent from '@mui/material/CardContent';
+import Card from '@mui/material/Card';
 import { useHistory } from 'react-router';
+import { useTheme } from '@mui/material/styles';
+import Box from '@mui/material/Box';
 import blank from '../../components/Theme/blank.png';
 import Updater from '../../utils/Updater';
 import UpdateDialog from '../../components/UpdateDialog';
 import AlertDialog from '../../components/AlertDialog';
 import { setActiveListItem, setUpdateChecked } from '../../reducers/MainReducer/Actions';
 import { MainContext } from '../../contexts/MainContextProvider';
-
-const useStyles = makeStyles((theme) => ({
-  heroContent: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(4, 0, 2),
-  },
-  content: {
-    flexGrow: 1,
-    overflow: 'auto',
-  },
-  container: {
-    paddingTop: theme.spacing(2),
-    paddingBottom: theme.spacing(2),
-  },
-}));
 
 const os = window.require('os');
 
@@ -43,7 +29,7 @@ const Home = () => {
   const [errorMessage, setErrorMessage] = useState(null);
   const [update, setUpdate] = useState(null);
 
-  const classes = useStyles();
+  const theme = useTheme();
   const history = useHistory();
 
   /**
@@ -95,7 +81,12 @@ const Home = () => {
 
   return (
     <div>
-      <div className={classes.heroContent}>
+      <Box
+        sx={{
+          backgroundColor: 'background.paper',
+          padding: theme.spacing(4, 0, 2),
+        }}
+      >
         <Container maxWidth="sm">
           <Typography variant="h4" align="center" color="textPrimary" gutterBottom>
             {language.cryptography}
@@ -104,8 +95,13 @@ const Home = () => {
             {language.cryptographySubtitle}
           </Typography>
         </Container>
-      </div>
-      <main className={classes.content}>
+      </Box>
+      <main
+        style={{
+          flexGrow: 1,
+          overflow: 'auto',
+        }}
+      >
         {update && update.updateAvailable ? (
           <UpdateDialog
             downloadUrl={update.updateUrl}
@@ -125,7 +121,13 @@ const Home = () => {
             ok={language.ok}
           />
         ) : null}
-        <Container maxWidth="lg" className={classes.container}>
+        <Container
+          maxWidth="lg"
+          sx={{
+            paddingTop: theme.spacing(2),
+            paddingBottom: theme.spacing(2),
+          }}
+        >
           <Grid container spacing={2}>
             <Grid item xs={12} md={6} lg={6}>
               <Card>
